Now using velocity in [ec/sec]

+++ firmware/sigmadrive/velocityprofiler/trapezoidprofile.h
@@ -5,6 +5,6 @@
 #include <vector>
 #include <stdint.h>
 
-std::vector<std::vector<float>> CalculateTrapezoidPoints(float Xi, float Xf, float Vin, float Vfin, float Vmax, float Amax, float Dmax, float Hz);
+std::vector<std::vector<int64_t>> CalculateTrapezoidPoints(float Xi, float Xf, float Vin, float Vfin, float Vmax, float Amax, float Dmax, float Hz);
 
 #endif
